% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_cv}
\alias{check_cv}
\title{Check TRN Assembly}
\usage{
check_cv(x, hybrid_split = "_")
}
\arguments{
\item{x}{A matrix containing cross-validation folds inside columns and
hybrid names in rows.}

\item{hybrid_split}{Character that separates parents in hybrid names.}
}
\value{
'success', if all checks were successful.
}
\description{
Check TRN Assembly
}
\keyword{internal}

