% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_trn}
\alias{check_trn}
\title{Check TRN Assembly}
\usage{
check_trn(x, hybrid_split = "_")
}
\arguments{
\item{x}{A matrix containing cross-validation folds inside columns and
hybrid names in rows.}

\item{hybrid_split}{Character that separates parents in hybrid names.}
}
\value{
Integer with the number of maternal and paternal training set lines.
}
\description{
Check TRN Assembly
}
\keyword{internal}

