% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_frequency.R
\name{compute_cf}
\alias{compute_cf}
\title{Call Frequency}
\usage{
compute_cf(x, output = c("markerNames", "markerCallfreq"), missing = "??",
  callThresh = NULL)
}
\arguments{
\item{x}{A matrix. Genotype names are stored in rows whereas marker names
are stored in columns.}

\item{output}{Character vector with output options.}

\item{missing}{Character vector providing the encoding of missing elements.}

\item{callThresh}{Numeric or complex vector specifying the call frequency
threshold.}
}
\value{
If \code{output} is "markerNames" a character vector with marker
 names that have passed the quality check will be returned. If \code{output}
 is "markerCallfreq" a numeric vector with the call frequency at each
 marker locus will be returned.
}
\description{
\code{compute_cf} returns results from call frequency check
}
\examples{
 # Load a matrix with SNP genotypes encoded as numeric values
 data(marker_numeric)

 # Compute the call frequency
 compute_cf(marker_numeric, output = "markerCallfreq", missing = NA_real_)

 # Retrieve the marker names of all loci with a call frequency >= 0.9.
 compute_cf(marker_numeric, output = "markerNames", missing = NA_real_,
            callThresh = 0.9)
}

