% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_snp_quality.R
\name{ensure_snp_quality}
\alias{ensure_snp_quality}
\title{Ensure SNP Quality}
\usage{
ensure_snp_quality(snp, callfreq_check = TRUE, callfreq_threshold = 0.95,
  maf_check = TRUE, maf_threshold = 0.05, any_missing = TRUE,
  missing_value = NA_character_, remove_duplicated = TRUE)
}
\arguments{
\item{snp}{A matrix. Genotype names are stored in rows whereas marker names
are stored in columns.}

\item{callfreq_check}{Logical scalar. Shall markers with a low call
frequency (*i.e.* a high number of missing marker genotypes) be removed?}

\item{callfreq_threshold}{Numeric scalar. Minimum level of non-missing
genotypes per locus.}

\item{maf_check}{Logical scalar. Shall markers with a low minor allele
frequency be removed?}

\item{maf_threshold}{Numeric scalar. Minimum frequency of the minor allele
at each locus.}

\item{any_missing}{Logical scalar. Does the input matrix \code{snp} contain
any missing values? If this is the case, they will be replaced with the
major genotype at this locus.}

\item{missing_value}{Specify the encoding of missing genotypes.}

\item{remove_duplicated}{Logical scalar. Should only unique marker loci be
returned?}
}
\value{
Depending on the choice of parameters \code{callfreq_check} and
 \code{maf_check}, respectively, \code{ensure_snp_quality} will return a
 matrix with marker genotypes that have passed important quality checks.
}
\description{
\code{ensure_snp_quality} applies quality checks to SNP data
}
\examples{
 # Load a matrix with SNP genotypes encoded as numeric values
 data(marker_numeric)

 # Add a duplicated marker locus to the data.
 snp <- marker_numeric
 snp <- cbind(snp, snp[, 1])

 # Return markers without missing values, a call frequency greater or equal
 # to 0.95 and a minor allele frequency greater or equal to 0.6. Finally,
 # remove all duplicated markers.
 ensure_snp_quality(snp, maf_threshold = 0.1, missing_value = NA_real_)

 # Load a matrix with SNP genotypes encoded as character values
 data(marker_character)

 # Return markers without missing values and a call frequency equal to or
 # greater than 0.9. Keep duplicated markers if present.
 ensure_snp_quality(marker_character, callfreq_threshold = 0.9,
                    maf_check = FALSE, missing_value = "??",
                    remove_duplicated = FALSE)
}
