% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hybrid_nms}
\alias{hybrid_nms}
\title{Hybrid names.}
\format{A character vector with hybrid names where the two parents are
 separated by an underscore '_'.}
\usage{
hybrid_nms
}
\description{
A vector containing the names of hybrids. It can be used for sampling a
cross-validation scheme.
}
\keyword{datasets}
