% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imp_snps}
\alias{imp_snps}
\title{Imputed, numeric SNP genotypes}
\format{A matrix with 10 rows and 20 columns. All genotypes have class
'numeric'. No missing values exist in the data set.}
\usage{
imp_snps
}
\description{
A dataset containing the genotypes of ten individuals for 20 loci. The
variables are as follows:
}
\keyword{datasets}

