% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute2.R, R/impute_eta.R
\name{impute_eta}
\alias{impute_eta}
\title{Imputation of ETA}
\usage{
impute_eta(x, y, geno = NULL, bglr_model)

impute_eta(x, y, geno = NULL, bglr_model)
}
\arguments{
\item{x}{A matrix with pedigree records.}

\item{y}{A matrix with SNP records.}

\item{geno}{NULL. Optional character vector with names corresponding to one
of the two parental pools.}

\item{bglr_model}{A character specifying the algorithm that shall be used \
when calling \code{BGLR()}.}

\item{z}{A matrix with transcriptomic BLUEs.}

\item{x}{A matrix with BLUES of the complete predictor.}

\item{y}{A matrix with BLUEs of the incomplete predictor.}

\item{geno}{NULL. Optional character vector with names corresponding to one
of the two parental pools.}

\item{bglr_model}{A character specifying the algorithm that shall be used \
when calling \code{BGLR()}.}
}
\value{
A list with model parameters used as input in the \code{BGLR()} call.

A list with model parameters used as input in the \code{BGLR()} call.
}
\description{
\code{impute2} returns a list with BGLR model parameters.

\code{impute_eta} returns a list with BGLR model parameters.
}
\examples{
data("hybrid_nms", "mrna", "imp_snps")
geno <- vapply(strsplit(hybrid_nms, split = "_"), FUN = "[[", 1,
               FUN.VALUE = character(1))
x <- imp_snps[rownames(imp_snps) \%in\% geno, ]
y <- mrna[rownames(mrna) \%in\% geno, ]
eta <- impute_eta(x = x, y = y, geno = geno, bglr_model = "BRR")
str(eta)
data("hybrid_nms", "mrna", "imp_snps")
geno <- vapply(strsplit(hybrid_nms, split = "_"), FUN = "[[", 1,
               FUN.VALUE = character(1))
x <- imp_snps[rownames(imp_snps) \%in\% geno, ]
y <- mrna[rownames(mrna) \%in\% geno, ]
eta <- impute_eta(x = x, y = y, geno = geno, bglr_model = "BRR")
str(eta)
}

