% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marker_character}
\alias{marker_character}
\title{SNP genotypes (character)}
\format{A matrix with 10 rows and 10 columns. All genotypes have class
'character'. The matrix contains missing values.}
\usage{
marker_character
}
\description{
A dataset containing the genotypes of ten individuals for ten loci. The
variables are as follows:
}
\keyword{datasets}
