% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mice_mrna}
\alias{mice_mrna}
\title{Mice transcriptomic data}
\format{A matrix with 96 rows and 400 columns.}
\usage{
mice_mrna
}
\description{
A matrix with transcriptomic features for 96 mice. The values were calculated
using a multivariate normal distribution where pedigree records in 'mice_ped'
specified the covariance matrix of the mRNA vectors.
}
\keyword{datasets}

