% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mice_ped}
\alias{mice_ped}
\title{Mice pedigree data}
\format{A matrix with 150 rows and 150 columns.}
\usage{
mice_ped
}
\description{
A numerator relationship matrix with pedigree records for 150 mice.
}
\keyword{datasets}
