% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mice_pheno}
\alias{mice_pheno}
\title{Mice phenotypic data}
\format{A data.frame with 150 rows and 2 columns.}
\usage{
mice_pheno
}
\description{
A data frame with obesity records for 150 mice taken from the 'BGLR' package.
}
\keyword{datasets}
