% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mice_snp}
\alias{mice_snp}
\title{Mice genomic relationship data}
\format{A matrix with 120 rows and 120 columns.}
\usage{
mice_snp
}
\description{
A genomic relationship matrix with records for 120 mice.
}
\keyword{datasets}

