% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mrna}
\alias{mrna}
\title{Transcriptomic data}
\format{A matrix with 44 rows and 500 columns.}
\usage{
mrna
}
\description{
A matrix containing the gene expression values for 44 genotypes and 500
mRNAs.
}
\keyword{datasets}
