% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pheno}
\alias{Pheno}
\title{Phenotypic data}
\format{A matrix with 532 rows and two columns.}
\usage{
Pheno
}
\description{
A matrix containing the agronomic data for 532 hybrids and two traits.
}
\keyword{datasets}
