% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_coding.R
\name{recode_snps}
\alias{recode_snps}
\title{Recode genotypic data}
\usage{
recode_snps(x, major, minor, major_coding, minor_coding, het_coding, na_coding)
}
\arguments{
\item{x}{A matrix. Genotype names are stored in rows whereas marker names
are stored in columns.}

\item{major}{Character vector with names of major genotype at each locus.}

\item{minor}{Character vector with names of minor genotype at each locus.}

\item{major_coding}{Scalar specifying how to recode the major genotype.}

\item{minor_coding}{Scalar specifying how to recode the minor genotype.}

\item{het_coding}{Scalar specifying how to recode heterozygous loci.}

\item{na_coding}{Scalar specifying how to recode missing values.}
}
\value{
A matrix with the same dimensions as \code{x} but with recoded SNP
 genotypes.
}
\description{
\code{recode_snps} recodes the elements of a SNP-matrix
}
\examples{
 # First, determine the major and minor allele at each locus of SNP-genotypes
 geno_lst <- compute_maf(marker_character, output = "genoList",
                         missing = "??", mafThresh = 0)
 major <- geno_lst[["major_allele"]]
 minor <- geno_lst[["minor_allele"]]

 # Recode the major allele as '2', the minor allele as '0' and heterozygotes
 # as '0'.
 recode_snps(marker_character, major = major, minor = minor,
             major_coding = 2, minor_coding = 0, het_coding = 1,
             na_coding = NA_real_)
}

