% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_cv.R
\name{sample_cv}
\alias{sample_cv}
\title{Cross-Validation Scheme Sampling}
\usage{
sample_cv(x, n_father, n_mother, n_hyb_trn, min_size, rounds, hybrid_split,
  progress = FALSE)
}
\arguments{
\item{x}{A vector with hybrid names.}

\item{n_father}{A numeric or integer scalar specifying the size of the
sampled paternal parents in the training set.}

\item{n_mother}{A numeric or integer scalar specifying the size of the
sampled maternal parents in the training set.}

\item{n_hyb_trn}{A numeric or integer scalar specifying the size of the
training set.}

\item{min_size}{A numeric or integer scalar specifying the minimum size of
each test set class (VS0, VS1, VS2).}

\item{rounds}{A numeric or integer speciyfing the number of cross-validation
runs to sample.}

\item{hybrid_split}{A character scalar specifying the character that
separates the maternal from the paternal parent in \code{x}.}

\item{progress}{A logical scalar. Set to 'TRUE' if a progress bar is desired.}
}
\value{
A matrix of dimensions \code{length(x)} by \code{rounds} with
 columns containing assignments of hybrids to either TRN, T0, T1 or T2.
}
\description{
\code{sample_cv} returns a matrix with cross-validation (CV) folds.
}
\examples{
data(hybrid_nms)
cv_mat <- sample_cv(hybrid_nms, n_mother = 39, n_father = 33,
                    n_hyb_trn = 200, min_size = 20, rounds = 20L,
                    hybrid_split = "_")
str(cv_mat)
}

