% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_eta.R
\name{complete_eta}
\alias{complete_eta}
\title{Construction of ETA with complete predictor data}
\usage{
complete_eta(x, geno, bglr_model)
}
\arguments{
\item{x}{A matrix with BLUES of the complete predictor.}

\item{geno}{NULL. Optional character vector with names corresponding to one
of the two parental pools.}

\item{bglr_model}{A character specifying the algorithm that shall be used \
when calling \code{BGLR()}.}
}
\value{
A list with model parameters used as input in the \code{BGLR()} call.
}
\description{
\code{complete_eta} returns a list with BGLR model parameters.
}
\examples{
data("hybrid_nms", "imp_snps")
geno <- vapply(strsplit(hybrid_nms, split = "_"), FUN = "[[", 1,
               FUN.VALUE = character(1))
x <- imp_snps[rownames(imp_snps) \%in\% geno, ]
eta <- complete_eta(x = x, geno = geno, bglr_model = "BRR")
str(eta)
}

