% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{compute_maf}
\alias{compute_maf}
\title{Minor Allele Frequency}
\usage{
compute_maf(x, output = c("markerNames", "markerMAF", "genoList"),
  missing = "??", mafThresh = 0)
}
\arguments{
\item{x}{A matrix. Genotype names are stored in rows whereas marker names
are stored in columns.}

\item{output}{Character vector with output options.}

\item{missing}{Character vector providing the encoding of missing elements.}

\item{mafThresh}{Numeric or complex vector specifying the minor allele
frequency threshold.}
}
\value{
If \code{output} is "markerNames" a character vector with marker
 names that have passed the quality check will be returned. If \code{output}
 is "markerMAF" a numeric vector with the minor allele frequency at each
 marker locus will be returned. If \code{output} is "genoList" a list of two
 elements will be returned, which provides the minor allele at each locus as
 well as the major allele. The speed of computations is considerably higher
 if the elements of \code{x} are supplied as character values since an
 internal class conversion will be applied otherwise.
}
\description{
\code{compute_maf} returns results from minor allele frequency check
}
\examples{
 # Load a matrix with SNP genotypes encoded as numeric values
 data(marker_numeric)

 # Return the names of all polymorphic markers.
 compute_maf(marker_numeric, output = "markerNames", missing = NA_real_,
             mafThresh = 0)

 # Set one locus to monomorphic
 marker_mono <- marker_numeric
 marker_mono[, 1] <- 1
 compute_maf(marker_numeric, output = "markerNames", missing = NA_real_,
             mafThresh = 0)

 # Load a matrix with SNP genotypes encoded as character values
 data(marker_character)

 # Return the minor allele frequency at each locus with MAF >= 0.05.
 compute_maf(marker_character, output = "markerMAF", missing = "??",
             mafThresh = 0.05)

 # Return the genotype of the major as well as of the minor allele at each
 # locus.
 compute_maf(marker_character, output = "genoList", missing = "??",
             mafThresh = 0)
}

