% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mice_snp}
\alias{mice_snp}
\title{Mice SNP data}
\format{A matrix with 120 rows and 1000 columns.}
\usage{
mice_snp
}
\description{
A feature matrix with SNP genotypes for 120 mice and 1000 markers.
}
\keyword{datasets}

